requirements.txt
Flask
easyocr
Pillow
opencv-python-headless
numpy
reportlab
pdf2image